# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Holding Today - a private place for bereaved parents
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /community/posts:
    get:
      operationId: getCommunityPosts
      tags: [community]
      summary: Recent posts in the shared room
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            type: string
            enum: [signs, story, advice]
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityPost"
    post:
      operationId: createCommunityPost
      tags: [community]
      summary: Share something with the room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityPostInput"
      responses:
        "201":
          description: Posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedCommunityPost"
  /community/posts/{id}:
    get:
      operationId: getCommunityPost
      tags: [community]
      summary: One post and its replies
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPostDetail"
        "404":
          description: No such post, or it has been removed
    delete:
      operationId: deleteCommunityPost
      tags: [community]
      summary: Take your own post down
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /community/posts/{id}/comments:
    post:
      operationId: createCommunityComment
      tags: [community]
      summary: Reply to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityCommentInput"
      responses:
        "201":
          description: Posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedCommunityComment"
  /community/comments/{id}:
    delete:
      operationId: deleteCommunityComment
      tags: [community]
      summary: Take your own reply down
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /community/reports:
    post:
      operationId: reportCommunityContent
      tags: [community]
      summary: Report a post or a reply
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityReportInput"
      responses:
        "201":
          description: Reported
  /community/screen-name:
    put:
      operationId: setScreenName
      tags: [community]
      summary: Choose the name you appear under in the shared room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenNameInput"
      responses:
        "200":
          description: Set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreenName"
  /community/moderation/reports:
    get:
      operationId: getCommunityReports
      tags: [community]
      summary: Open reports. Moderators only.
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityReportEntry"
        "403":
          description: Not a moderator
  /community/moderation/hide:
    post:
      operationId: hideCommunityContent
      tags: [community]
      summary: Hide or restore a post or reply. Moderators only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HideInput"
      responses:
        "204":
          description: Done
        "403":
          description: Not a moderator
...
components:
  schemas:
...
    CommunityPost:
      type: object
      properties:
        id:
          type: integer
        screenName:
          type: string
        kind:
          type: string
          enum: [signs, story, advice]
        title:
          type: string
        body:
          type: string
        commentCount:
          type: integer
        isMine:
          type: boolean
          description: Whether the signed-in reader wrote it, so they can remove it.
        createdAt:
          type: string
      required: [id, screenName, kind, title, body, commentCount, isMine, createdAt]
    CommunityPostInput:
      type: object
      properties:
        kind:
          type: string
          enum: [signs, story, advice]
        title:
          type: string
        body:
          type: string
        screenName:
          type: string
          nullable: true
          description: >-
            Leave blank and one is assigned. Only used the first time; after
            that the account keeps the name it already has.
      required: [kind, title, body]
    CreatedCommunityPost:
      allOf:
        - $ref: "#/components/schemas/CommunityPost"
        - type: object
          properties:
            crisisPrompt:
              type: boolean
              description: >-
                True when the post reads like the author may be in danger. The
                post is published either way; this asks the client to put the
                crisis numbers in front of them.
          required: [crisisPrompt]
    CommunityComment:
      type: object
      properties:
        id:
          type: integer
        screenName:
          type: string
        body:
          type: string
        isMine:
          type: boolean
        createdAt:
          type: string
      required: [id, screenName, body, isMine, createdAt]
    CreatedCommunityComment:
      allOf:
        - $ref: "#/components/schemas/CommunityComment"
        - type: object
          properties:
            crisisPrompt:
              type: boolean
          required: [crisisPrompt]
    CommunityCommentInput:
      type: object
      properties:
        body:
          type: string
      required: [body]
    CommunityPostDetail:
      allOf:
        - $ref: "#/components/schemas/CommunityPost"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/CommunityComment"
          required: [comments]
    CommunityReportInput:
      type: object
      properties:
        targetType:
          type: string
          enum: [post, comment]
        targetId:
          type: integer
        reason:
          type: string
          enum: [cruel, spam, selling, graphic, personal_information, worried_about_them, other]
        note:
          type: string
          nullable: true
      required: [targetType, targetId, reason]
    CommunityReportEntry:
      type: object
      properties:
        id:
          type: integer
        targetType:
          type: string
        targetId:
          type: integer
        reason:
          type: string
        note:
          type: string
          nullable: true
        createdAt:
          type: string
        targetBody:
          type: string
          nullable: true
        targetScreenName:
          type: string
          nullable: true
        targetHidden:
          type: boolean
        reportCount:
          type: integer
          description: How many separate people have reported this same thing.
      required: [id, targetType, targetId, reason, createdAt, targetHidden, reportCount]
    HideInput:
      type: object
      properties:
        targetType:
          type: string
          enum: [post, comment]
        targetId:
          type: integer
        hidden:
          type: boolean
        reason:
          type: string
          nullable: true
      required: [targetType, targetId, hidden]
    ScreenName:
      type: object
      properties:
        screenName:
          type: string
      required: [screenName]
    ScreenNameInput:
      type: object
      properties:
        screenName:
          type: string
          nullable: true
          description: Blank asks the server to assign one.
      required: []
...